refactor(devtools): single source of truth for shared protocol types
The VNode tree node, component detail entry, and render event shapes were
declared three times with identical fields: in the browser extension, in the
devtools UI, and in the kit client bridge.

Move them into the kit protocol module (DevtoolsVNodeTreeNode in protocol/vnode,
DevtoolsRenderEvent in protocol/perf, DevtoolsComponentDetailEntry in
protocol/hooks). The client bridge, the extension, and the UI now all import
them from @qwik.dev/devtools/kit. No public API change: the same names are still
exported, just from the protocol module.
